% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c19model.R
\name{c19model}
\alias{c19model}
\title{c19model Class constructor / factory / generator}
\usage{
c19model(inputs, level)
}
\arguments{
\item{inputs}{the inputs object created with get_augmented_inputs}

\item{level}{either "state" or "country".}
}
\value{
a c19model object.
}
\description{
This is a template class constructor for models based on the c19model class.
}
