% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_c19_2.R
\name{c19model_amshi2s}
\alias{c19model_amshi2s}
\title{C19 Model 2 Function}
\usage{
c19model_amshi2s(inputs, level)
}
\arguments{
\item{inputs}{inputs object created with the get_augmented_inputs () and prepare_inputs() functions}

\item{level}{use "state" as default.}
}
\value{
a c19model object
}
\description{
C19 Model 2 Function
}
\examples{
\dontrun{
model = get_augmented_inputs(...) \%>\%
 prepare_inputs(...) \%>\%
 c19model_amshi2s(., level = "state")
}

}
