% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c19model_03_calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate}
\usage{
calibrate(
  model,
  location_ids,
  n_lhs = 2,
  minR0 = 2,
  maxR0 = 4,
  parallel_mode = "FORK",
  cores,
  test_run = F,
  scenarios = NULL
)
}
\arguments{
\item{model}{a model of the c19model class}

\item{location_ids}{a vector of location ids}

\item{n_lhs}{the number of samples to use}

\item{minR0}{the min R0 that will be used on the calibration process by the set_calibrated_k function}

\item{maxR0}{the max R0 that will be used on the calibration process by the set_calibrated_k function}

\item{parallel_mode}{either FORK or PSOCK. Use PSOCK on windows and FORK on UNIX systems.}

\item{cores}{number of CPU cores to use.}
}
\value{
a calibrated model of the c19model class
}
\description{
Calibrates the c19model
}
