% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c19model_03_calibrate.R
\name{calibrate_initiation_num}
\alias{calibrate_initiation_num}
\title{Calibrates the Initiation Number}
\usage{
calibrate_initiation_num(
  run_id,
  model,
  tolerance = 0.01,
  max_runs = 10,
  initiation.num = 1e-04,
  calibration_day
)
}
\arguments{
\item{run_id}{The RunID (integer)}

\item{model}{the c19model object}

\item{tolerance}{percent error to tolerate}

\item{max_runs}{maximum number of runs to improve initiation number}

\item{initiation.num}{initial initiation number}

\item{calibration_day}{day (integer) to use in the process of calibrating the initiation number}
}
\value{
results dataset from the initiation number.
}
\description{
Calibrates the unknown number of infected people at the beginning of the epidemic.
}
