% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c19model_03_calibrate.R
\name{calibrate_params_lhs}
\alias{calibrate_params_lhs}
\title{Runs a Single Calibration Run and compute fit stats}
\usage{
calibrate_params_lhs(run_id, model, calibration_day)
}
\arguments{
\item{run_id}{The run ID to use}

\item{model}{a model object of the c19model class}
}
\value{
a results dataframe including fit statistics
}
\description{
Runs a Single Calibration Run and compute fit stats
}
