% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c19model_01_get_inputs.R
\name{compute_agregated_stats}
\alias{compute_agregated_stats}
\title{computes additional aggregate statistics}
\usage{
compute_agregated_stats(df)
}
\description{
computes additional aggregate statistics
}
