% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c19model_05_compute_outputs.R
\name{compute_augmented_outputs}
\alias{compute_augmented_outputs}
\title{Computes Additional Outputs}
\usage{
compute_augmented_outputs(
  model,
  compute_econ_outputs = T,
  calibration = T,
  ...
)
}
\arguments{
\item{model}{a c19model object}

\item{compute_econ_outputs}{TRUE if it should compute additional economic outputs}

\item{calibration}{TRUE if computing results from calibration}
}
\value{
the model object including augmented outputs
}
\description{
Computes Additional Outputs
}
