% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c19model_04_run_experiments.R
\name{evaluate_experiment}
\alias{evaluate_experiment}
\title{Evaluate Experiments (in Parallel)}
\usage{
evaluate_experiment(
  model,
  sim_end_date,
  runs = 1:nrow(model$future_experimental_design),
  n_cores = parallel::detectCores() - 2,
  parallel_mode = "PSOCK",
  solver = "lsoda",
  write_csv = F,
  test_run = F
)
}
\arguments{
\item{model}{The c19model object}

\item{sim_end_date}{Simulation end date (in the yyyy-mm-dd format, e.g." "2021-01-20")}

\item{runs}{Vector of Runs to perform (should correspond to ExperimentIDs defined in the future_experimental_design data.frame)}

\item{n_cores}{Number of CPU cores to use, defaults to total number of cores - 2}

\item{parallel_mode}{Parallel Cluster type to use. Either "PSOCK" or "FORK". Windows users, use "PSOCK", otherwise, use "FORK"}

\item{solver}{"lsoda" by default, but could use any deSolve-compatible solver.}

\item{write_csv}{TRUE if you want to save preliminary results to an ./output folder. Make sure this folder exists and is empty.}
}
\value{
a data.frame with simulation results
}
\description{
This function evaluates experiments in parallel, and can be used to scale model runs vertically and horizontally. You can first define your experiments using the \code{set_parameter} and \code{set_experimental_design} functions. Then, you can save your model object and send it to different servers, if you want. Then, use this function to evaluate the experiments in parallel, while saving results to the ./outputs folder. These results can be visualized before your complete experiments are ready.
}
