% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c19model.R
\name{filter_inputs}
\alias{filter_inputs}
\title{Filter Inputs for simulation}
\usage{
filter_inputs(inputs, location_ids, level = "state")
}
\arguments{
\item{inputs}{an inputs list}

\item{location_ids}{the location ids to filter}

\item{level}{the model level to use. Currently this is not used.}
}
\value{
the same list
}
\description{
Filters the location, economicinputs, timeseries and compartment objects of the selected location_ids
}
