% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_generate_webtool_data.R
\name{generate_webtool_data}
\alias{generate_webtool_data}
\title{Generate WebTool Data}
\usage{
generate_webtool_data(
  model,
  scenario_runs,
  folder = "./webdata/",
  update_past_data = F,
  update_scenarios = F,
  append = F
)
}
\arguments{
\item{model}{epidemiological model to use}

\item{scenario_runs}{data.frame containing the results from the scenario runs.}

\item{folder}{folder in which to save csv files}

\item{update_past_data}{TRUE if you want to update past data}

\item{update_scenarios}{TRUE if you want to update simulation results}

\item{append}{TRUE if you want to append results to the csv files}
}
\description{
Generates the csv files that are the source of all data at the webtool.
}
