% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c19model_01_get_inputs.R
\name{get_augmented_inputs}
\alias{get_augmented_inputs}
\title{Get Augmented Inputs}
\usage{
get_augmented_inputs(
  inputs_path = "./data/inputs",
  inputs_filename = "input_dataset.xlsx",
  econ_inputs_filename = "econ_inputs.xlsx",
  unemployment_data_filename = "unemployment_data.xlsx",
  write_to_server = F,
  read_from_web = T,
  get_countries = F,
  start_date = "2020-03-01",
  calibration_date = "2020-04-07",
  get_unemployment_data = F
)
}
\arguments{
\item{write_to_server}{if TRUE, writes the augmented inputs file to a server}

\item{read_from_web}{if TRUE, gets timeseries data from the web, for US states. If you use FALSE, make sure to have a locationtimeseries object in your inputs file.}

\item{start_date}{The data in which simulation will begin. This data influences the timeseries gathering, and calibration.}

\item{calibration_date}{Date when the initiation number will be calibrated. You should observe at least one death in each location for that date.}

\item{get_unemployment_data}{True if user wants to collect and gather unemployment data from pdf data.}

\item{filename}{standardized inputs xlsx file}
}
\value{
augm_inputs list object, which will be used throughout the package.
}
\description{
This function reads an inputs spreadsheet and returns an augmented list of data.frame. The augmented inputs object is used to create models and contain a set of data.frames and objects that represent the initial structure of the model.
}
\examples{
\dontrun{
augm_inputs = get_augmented_inputs("./data/inputs/input_dataset_combined.xlsx", 
                                  start_date = "2020-03-01", 
                                 calibration_date = "2020-04-13")
}

}
