% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c19model_05_compute_outputs.R
\name{get_best_fit_model}
\alias{get_best_fit_model}
\title{Get Best Fit Runs}
\usage{
get_best_fit_model(model, n = 1)
}
\arguments{
\item{model}{A calibrated c19model object}

\item{n}{The number of runs to get.}
}
\value{
a model object with filtered runs. The cal_results and the cali_augm_results tables are filtered.
}
\description{
This function selects the n best-fitting calibration runs for each location in a calibrated model.
}
