% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c19model_01_get_inputs.R
\name{get_inputs_object}
\alias{get_inputs_object}
\title{Reads the inputs spreadsheet and saves it as a list.}
\usage{
get_inputs_object(filename = "inputs_data.xlsx")
}
\arguments{
\item{filename}{path to data-dictionary.xls file. This file should always be placed inside a /data folder.}
}
\value{
a list containing each tab of the data dictionary spreadsheet as a data.frame
}
\description{
This function reads an standardized excel spreadsheet and creates a list containing all the naming conventions that will be used to give variables a user-friendly name. You should only read a standardized data-dictionary.xlsx file. To get one, run create_fam_app_folder() and navigate to the /data folder.
}
