% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c19model_03_calibrate.R
\name{get_ode_fit}
\alias{get_ode_fit}
\title{Get ODE fit statistics}
\usage{
get_ode_fit(
  model_timeseries,
  actual_timeseries,
  model_outcomes = c("Deaths", "CurrentlyHospitalized", "PositiveTests"),
  data_outcomes = c("DeathsMovingAverage", "CurrentlyHospitalizedMovingAverage",
    "PositiveTestsMovingAverage"),
  outcomes_weights = c(1, 1e-04, 1e-05),
  time_discount_factor = 0.1
)
}
\arguments{
\item{model_timeseries}{model resulta dataframe}

\item{actual_timeseries}{actual data dataframe}

\item{model_outcomes}{vector of outcomes names to be used}

\item{data_outcomes}{vector of data outcome names to be used}

\item{outcomes_weights}{weights to be used when computing the weighted RMSE}

\item{time_discount_factor}{not used for now, but could be used in the future}
}
\value{
a list with fit statistics
}
\description{
Compares the ODE results with a set of timeseries. Computes a Weighted Root Mean Squared Error based on pre-specified weights.
}
