% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c19model_05_compute_outputs.R
\name{get_prcc}
\alias{get_prcc}
\title{Gets Partial Rank Correlation Statistics}
\usage{
get_prcc(model, location_id, outcome_variable)
}
\arguments{
\item{model}{model object}

\item{location_id}{A location id}

\item{outcome_variable}{THe outcome variable to use}
}
\value{
a dataframe with results from the PRCC analysis
}
\description{
Gets Partial Rank Correlation Statistics
}
