% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c19model_05_compute_outputs.R
\name{plot_param_fit}
\alias{plot_param_fit}
\title{Plots scatter plots showing if any of the parameters is driving model bias}
\usage{
plot_param_fit(calibrated_model, param, fit.stat = "fit.DeathsPercentBias")
}
\arguments{
\item{calibrated_model}{The Calibrated Model}

\item{param}{parameter name}

\item{fit.stat}{The fit statistic to display in the plot}
}
\value{
a plot
}
\description{
Plots scatter plots showing if any of the parameters is driving model bias
}
