% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c19model_02_prepare_inputs.R
\name{prepare_inputs}
\alias{prepare_inputs}
\title{Prepare Inputs}
\usage{
prepare_inputs(augm_inputs)
}
\arguments{
\item{augm_inputs}{inputs object created with the get_augmented_inputs function.}
}
\value{
The transformed inputs object
}
\description{
Prepares inputs for the simulation model. This includes converting tables to matrices, computing contact matrices, defining calibration dates and running exponential regressions.
}
\examples{
\dontrun{
model_inputs = augm_inputs \%>\% prepare_inputs()
}

}
