% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c19model_04_run_model.R
\name{run_all_scenarios}
\alias{run_all_scenarios}
\title{Runs All (Pre-Specified) Scenarios}
\usage{
run_all_scenarios(
  model,
  portfolio_ids1,
  location_ids,
  start_dates,
  duration1,
  portfolio_ids2,
  portfolio_combination_mode,
  uncertainty_mode,
  uncertainty_runs,
  sim_end_date,
  parallel_mode,
  cores,
  solver,
  test_run = F,
  ...
)
}
\arguments{
\item{model}{a model object}

\item{portfolio_ids1}{a vector of portfolio ids to change to}

\item{location_ids}{A vector of location ids}

\item{start_dates}{a vector of dates to use}

\item{duration1}{the duration of portfolio 1}

\item{portfolio_ids2}{a vector of portfolios to change to}

\item{portfolio_combination_mode}{use "new_normal"}

\item{uncertainty_mode}{use "best_fitting", but we could define other types of uncertainty analysis}

\item{uncertainty_runs}{number of runs to simulate.}

\item{sim_end_date}{the end date of the simulation (passed as a string aaaa-mm-dd)}

\item{parallel_mode}{either FORK or PSOCK}

\item{cores}{number of CPU cores to use.}
}
\value{
a list including the results and the future scenarios simulated.
}
\description{
Runs All (Pre-Specified) Scenarios
}
