% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c19model_04_run_model.R
\name{run_calibration_in_parallel}
\alias{run_calibration_in_parallel}
\title{Runs calibration runs in parallel}
\usage{
run_calibration_in_parallel(
  model,
  run_mode = "scenarios",
  parallel_mode,
  cores,
  ...
)
}
\arguments{
\item{model}{The model object}

\item{run_mode}{either "scenarios", "init.num" or "calibration_lhs". We start the calibration process with init.num, then move to "calibration_lhs" and then "scenarios".}

\item{parallel_mode}{either "FORK" or "FORK". "PSOCK" works on windows machines}

\item{cores}{number of CPU cores to use.}

\item{...}{Additional parameters to be passed to each model run function.}
}
\value{
A data.frame with simulation results.
}
\description{
Runs calibration runs in parallel
}
