% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c19model_04_run_experiments.R
\name{run_experiment}
\alias{run_experiment}
\title{Runs a Single Experiment}
\usage{
run_experiment(experiment_id, model, sim_end_date, solver = "lsoda", write_csv)
}
\arguments{
\item{experiment_id}{The ExperimentID}

\item{model}{The c19model object}

\item{sim_end_date}{Simulation end date (in the yyyy-mm-dd format, e.g." "2021-01-20")}

\item{solver}{"lsoda" by default, but could use any deSolve-compatible solver.}

\item{...}{Additional parameters to be passed to the inner function.}
}
\value{
data.frame with simulation results.
}
\description{
Runs a single Experiment defined by the set_experimental_design function. The experiment_id corresponds to the ExperimentID collumn in the future_experimental_design data.frame
}
