% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c19model_04_run_model.R
\name{run_single_future_run}
\alias{run_single_future_run}
\title{Runs a Single Future Run}
\usage{
run_single_future_run(run_id, model, solver, ...)
}
\arguments{
\item{run_id}{The RunID}

\item{model}{The Model Object}
}
\value{
the results dataframe
}
\description{
Runs a Single Future Run
}
