% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c19model_04_run_experiments.R
\name{set_experimental_design}
\alias{set_experimental_design}
\title{Set Experimental Design}
\usage{
set_experimental_design(
  model,
  n_new_lhs,
  convert_lhs_to_grid = F,
  lhs_to_grid_midpoints = 0
)
}
\arguments{
\item{model}{The c19model object}

\item{n_new_lhs}{The number of points in the Latin Hypercube Sample to be created.}
}
\value{
The model object including a future_experimental_design data.frame.
}
\description{
Creates the future_experimental_design data.frame based on the paramers defined by the set_parameter functions. The experimental design created by this function is useful to run a typical RDM analysis where each policy is evaluated across a LHS of deep uncertainties. To achieve that, define each policy lever as a grid parameter, and each uncertainty as an "lhs" uncertainty.
}
