% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c19model.R
\name{set_lhs_scenarios}
\alias{set_lhs_scenarios}
\title{Sets the latin hypercube sample}
\usage{
set_lhs_scenarios(
  model,
  n_lhs = 100,
  seed = 1,
  lhs_mode = "calibration",
  probabilistic = T
)
}
\arguments{
\item{model}{A model of the c19model class}

\item{n_lhs}{Number of samples}

\item{seed}{Seed to set, if any}

\item{lhs_mode}{Should be "calibration" for now, but we may want to define different parameter bounds for different tasks.}

\item{probabilistic}{If TRUE, uses the probability distribution specified. Uses an uniform distribution otherwise}
}
\value{
The model object including a scenarios object.
}
\description{
Sets the latin hypercube sample
}
