% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c19model_03_compare_models.R
\name{set_model_comparison_experimental_design}
\alias{set_model_comparison_experimental_design}
\title{Set Calibration Experimental Design}
\usage{
set_model_comparison_experimental_design(model, location_ids, inputs_path)
}
\arguments{
\item{model}{The c19model object}
}
\value{
The model object including a calibration_experimental_design data.frame.
}
\description{
This function is similar to the set_experimental_design function, but it assumes
you want to calibrate the same model with different parameters. These parameters
usually change the model structure.
}
