% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c19model_04_run_experiments.R
\name{set_parameter}
\alias{set_parameter}
\title{Set Parameter}
\usage{
set_parameter(model, parameter_name, experimental_design, values, min, max)
}
\arguments{
\item{model}{the c19model object that has already been calibrated.}

\item{parameter_name}{charachter string defining the parameter name.}

\item{experimental_design}{Either "grid" or "lhs" Use lhs if you want to create a Latin Hypercube Sample within the min and max bounds you provided. Use Grid}

\item{values}{use when experimental_design = "grid". This should be a numeric vector including the values to be included in a grid experimental design.}

\item{min}{use when experimental_design = "lhs". This should bea numeric value indicating the minimum bound in the Latin Hypercube sample.}

\item{max}{use when experimental_design = "lhs". This should bea numeric value indicating the minimum bound in the Latin Hypercube sample.}
}
\value{
a c19 model object including the defined experimental parameters.
}
\description{
This functions constructs the experimental_parameter object, and appends experimental parameters that will be visible inside the model in the future.
Experimental parameters can be either uncertainties or decision levers.
Every parameter defined in this function can be accessed within the model by using \code{experimental_parameters$param_name}.
}
\details{
Experimental Parameters are not visible during calibration.
}
