% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c19model_04_run_model.R
\name{set_up_cluster}
\alias{set_up_cluster}
\title{Setup Cluster for Parallel Simulations}
\usage{
set_up_cluster(cluster, model, parallel_mode)
}
\arguments{
\item{cluster}{A cluster object created with makeCluster}

\item{model}{the model objecto to use}

\item{parallel_mode}{Either "FORK" or "PSOCK"}
}
\description{
Exports relevant objects in case we are using a FORK parallel cluster.
}
