% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_misc.R
\name{setup_c19_folder}
\alias{setup_c19_folder}
\title{Set up a COID-19 Folder}
\usage{
setup_c19_folder(path, overwrite = F)
}
\arguments{
\item{path}{The path where you want to create the folder. If the path is not a directory, the function will create a /c19analysis inside this path. If missing, defaults to the current working directory.}

\item{overwrite}{If TRUE, will overwrite an existing folder.}
}
\description{
Use this function to create a folder structure for the model.
}
