% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c19model_04_run_model.R
\name{simulate_single_scenario}
\alias{simulate_single_scenario}
\title{Simulates Single Scenario Building an intervention schedule}
\usage{
simulate_single_scenario(
  model,
  state_id,
  run_id,
  sim_end_date = "2020-12-31",
  npi1 = 6,
  npi2 = 6,
  npi3 = 1,
  date1 = "2020-05-01",
  date2 = "2020-06-01",
  date3 = "2020-07-01",
  solver = "lsoda",
  ...
)
}
\arguments{
\item{model}{the model object}

\item{state_id}{location id}

\item{run_id}{run_id to use}

\item{sim_end_date}{when to stop simulation}

\item{npi1}{first NPI}

\item{npi2}{second NPI}

\item{npi3}{third NPI}

\item{date1}{date to transition to the first NPI}

\item{date2}{date to transition to the second NPI}

\item{date3}{date to transition to the third NPI}
}
\value{
model object containing simulation results
}
\description{
Simulates Single Scenario Building an intervention schedule
}
