% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c19model_04_run_model.R
\name{solve_model}
\alias{solve_model}
\title{Solves the c19model}
\usage{
solve_model(model, run_id, solver = "lsoda", ...)
}
\arguments{
\item{model}{a model object of the c19model class}

\item{run_id}{the run id}

\item{solver}{lsoda is the preferable option}
}
\value{
a results dataframe
}
\description{
Solves the c19model
}
