% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c19model.R
\name{uniformize_inputs_level}
\alias{uniformize_inputs_level}
\title{Uniformize inputs}
\usage{
uniformize_inputs_level(inputs, level)
}
\arguments{
\item{inputs}{an inputs list}

\item{level}{the level to use (either state, city or country)}
}
\value{
the inputs list
}
\description{
The inputs spreadsheet contains inputs on the country and state level. This function unformizes the input object so all dataframes and objects have a consistent name.
}
