% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_update_and_calibrate.R
\name{update_and_calibrate_covid19model}
\alias{update_and_calibrate_covid19model}
\title{Calibrates and Runs The Model for a set of Locations}
\usage{
update_and_calibrate_covid19model(
  calibrated_model,
  location_ids,
  calibrate = T,
  n_calibration,
  n_uncertainty,
  portfolio_combination_mode,
  portfolio_ids1,
  start_dates,
  duration1,
  portfolio_ids2,
  start_date,
  sim_end_date,
  augm_inputs_path,
  parallel_mode = "FORK",
  cores,
  append = T,
  test_run = F,
  solver = "lsoda",
  ...
)
}
\arguments{
\item{calibrated_model}{An optional calibrated model object}

\item{location_ids}{The location ids to run scenarios}

\item{calibrate}{TRUE if you want to calbrate the model}

\item{n_calibration}{The number of samples to use in the calibration process}

\item{n_uncertainty}{The number of samples to use when analysing uncertainties}

\item{portfolio_combination_mode}{use "new_normal" to revert back to new normal protfolios}

\item{portfolio_ids1}{first portfolio to use}

\item{start_dates}{when to change to each portfolio}

\item{duration1}{number of days to stay in first portfolio}

\item{portfolio_ids2}{second portfolio to use}

\item{start_date}{Start Date of simulation.}

\item{sim_end_date}{End date of simulation aaaa-mm-dd}

\item{augm_inputs_path}{Path to the augmented inputs object that will be used.}

\item{parallel_mode}{FORK or PSOCK}

\item{cores}{number of CPU cores to use.}

\item{append}{whether or not to append}
}
\value{
The scenario_runs object
}
\description{
Calibrates and Runs The Model for a set of Locations
}
