% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_update_daily_timeseries.R
\name{update_daily_timeseries}
\alias{update_daily_timeseries}
\title{Update Daily Timeseries Data}
\usage{
update_daily_timeseries(
  inputs = "./data/inputs",
  datafolder = "../../covid19-cloudsearch-tool/data_files/"
)
}
\arguments{
\item{inputs}{path to inputs_dataset spreadsheet}

\item{datafolder}{path in which to save the csv files}
}
\description{
Updates csv files that are used by the web tool api, runs all tests and save the test logs.
}
